requirements.txt
aiokafka==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.14.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
async-property==0.2.2 ; python_version >= "3.12" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.2.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
coverage[toml]==7.6.8 ; python_version >= "3.12" and python_version < "4.0"
cryptography==44.0.0 ; python_version >= "3.12" and python_version < "4.0"
dependency-injector==4.43.0 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.15 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.5 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.56.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.68.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
jwcrypto==1.5.6 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.6 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-jaeger-proto-grpc==1.21.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-jaeger-thrift==1.21.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-jaeger==1.21.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-common==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-grpc==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-http==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-prometheus==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-aiokafka==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-proto==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-sdk==1.28.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-semantic-conventions==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-util-http==0.49b2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-client==0.21.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.27.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.10.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pytest-asyncio==0.24.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-cov==6.0.0 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.3 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-keycloak==4.7.3 ; python_version >= "3.12" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy-utils==0.41.2 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy[asyncio]==2.0.36 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.41.3 ; python_version >= "3.12" and python_version < "4.0"
structlog==24.4.0 ; python_version >= "3.12" and python_version < "4.0"
thrift==0.21.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.32.1 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
